Relocated some functionality
